EWM-222. Onboarding import wallet
