--- lib/feature/add_seed/import_wallet/import_wallet.dart
@@ -1,5 +1,0 @@
-export 'data/import_wallet_data.dart';
-export 'import_wallet_model.dart';
-export 'import_wallet_page.dart';
-export 'import_wallet_widget_model.dart';
-export 'view/view.dart';
