--- lib/feature/add_seed/import_wallet/view/view.dart
@@ -1,2 +1,0 @@
-export '../import_wallet_page.dart';
-export 'import_wallet_view.dart';
